# Optimizing classification functions