Thanks to @wasade 's insightful advices in issue #50 , I started to optimize the functions `find_rank`, `get_lineage` and `find_lca` accordingly. One advice I adopted is to replace `try...except...` with `dict.get` when looping up the hierarchies. This improves the performance quite a bit (see below). By design, there is a good chance that some taxa don't have rank assignments, so the error handling could cost some time. I didn't change the first `try...except...` before loop though, because the subject should be found in the tree, otherwise it is not classifiable. But once it is in the tree, the rest of hierarchies are guaranteed to be intact, and `tree[taxon]` won't cause any error. I also applied method reference and other tricks.

Benchmarks on 100,000 Bowtie2 SAM lines against the WoL database:

`find_rank`
species level: 64.3 ms ± 343 µs => 54.5 ms ± 119 µs
genus level: 75.6 ms ± 213 µs => 67.1 ms ± 203 µs
phylum level: 115 ms ± 192 µs => 111 ms ± 1.75 ms

`get_lineage`:
170 ms ± 592 µs => 133 ms ± 308 µs

`find_lca`:
189 ms ± 809 µs => 184 ms ± 1.55 ms

The `find_lca` will be further optimized in another PR.