@qiyunzhu 2020-06-28
The benchmarks for LCA functions are in issue #65 .