@wasade 2020-06-28
What's the reason for using a custom tree object rather than the thoroughly used and tested one in `skbio`?