@wasade 2020-06-28
Is there a reason `get_lineage` can't be computed and cached a single time? The number of names shouldnt be that large -- can't this effectively be expressed as `{'name_being_looked_up': ['some', 'lineage', 'name_being_looked_up']}`?