@qiyunzhu 2020-06-29
Hi @wasade :

1. The "tree" in Woltka is a much simpler structure. It is but a dictionary of current taxon to parent taxon. This structure is the same as taxdump. The advanced features of skbio's TreeNode are not necessary for the goal of Woltka.

2. It is true that `get_lineage` can be cached. The reason I haven't done so is because this function is only used in two places: one is `find_lca`, which is called for every set of subjects, and is bottlenecking. I already added cache for it so problem is solved. The other place is the final printing of lineage strings. It is not a bottle necking step because the number of rows in a feature table can't be very huge. Plus, all features are unique, so caching this function likely won't help.