@wasade 2020-06-30
Okay, but it seems like a bunch of the functionality is being reimplemented which is not ideal. And ultimately, if there are performance improvements, it would be helpful to put those into a common package for general benefit.